cromo: fix linter issues
Fixed various issues reported by cpplint.py when run on
src/platform2/cromo.

BUG=None
TEST=cpplint.py `find cromo/* | grep "\.cc\|\.h"`
     FEATURES=test emerge-link platform2

Reviewed-on: https://chromium-review.googlesource.com/210529
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>

+++ cromo/main.cc
@@ -73,7 +73,7 @@ static void exit_main_loop(void) {
 }
 
 static gboolean do_signal(void* arg) {
-  long sig = reinterpret_cast<long>(arg);
+  intptr_t sig = reinterpret_cast<intptr_t>(arg);
   LOG(INFO) << "Signal: " << sig;
 
   if (sig == SIGTERM || sig == SIGINT) {
